Move extractDomainIANASuffix from core to policy
It is currently in use only there. That's why
I also changed it to unexported.
